rtc: recycle id when unloading a rtc driver
When calling rtc_device_unregister, we are not freeing the id used by the
driver.
So when doing a unload/load cycle for a RTC driver (e.g. rmmod rtc_cmos
&& modprobe rtc_cmos), its id is incremented by one. As a consequence,
we no longer have neither an rtc0 driver nor a /proc/driver/rtc (as it
only exists for the first driver).

Signed-off-by: Vincent Palatin <vpalatin@chromium.org>

BUG=None
TEST=on x86 platform, rmmod rtc_cmos && modprobe rtc_cmos
and see the new rtc driver is still /sys/class/rtc/rtc0 and
/proc/driver/rtc still exists.

Change-Id: I1b60e25611d30b46a3744d38dc34213db6e21b72
Reviewed-on: https://gerrit.chromium.org/gerrit/29103
Reviewed-by: Derek Basehore <dbasehore@chromium.org>
Tested-by: Derek Basehore <dbasehore@chromium.org>
Tested-by: Vincent Palatin <vpalatin@chromium.org>
Reviewed-by: Olof Johansson <olofj@chromium.org>
Commit-Ready: Vincent Palatin <vpalatin@chromium.org>