mmc: dw_mmc: apply dynamic CardReadThreshold and FIFO related
This patch adds to apply dynamic CDTHRCTL(Card Read Threshold
Control) and FIFO related(Burst Size, FIFO Watermark)
Most card such as eMMC/mmc/sd card typically support block sizes
that are specified in the card or fixed to 512byte.
For SDIO cards, the block size is variable and can be chosen by
the application.

In order to use Card Read Threshold feature effectively and to
guarantee that the card clock does not stop. because of a FIFO
Full condition in the middle of block of data being read from
the Card.

The current driver code was setup for eMMC/mmc/sd that has fixed
512 block size. So this patch is for SDIO and specially CDTHRCTL
enable is mandatory for UHS_SDR104 mode

Change-Id: I1c1917b81a082839506c36e28ae5b4601a24e9ce
Signed-off-by: Hosung Kim <hosung0.kim@samsung.com>

Conflicts:

	drivers/mmc/host/dw_mmc.c