video: s3c-fb: rewrite all registers when they fail to update
Sometimes the s3c shadowed regsisters never seem to update when vsync
occurs.  When this happens, rewrite all the registers before waiting
for the next vsync.  If the shadow registers do not update after 100
tries, print an error and continue.  Worst case, it will page fault
if the old registers are still being used, which is better than
hanging the sync timeline.

Change-Id: I24ed94487dd288aaa7004cc68247a3a82f28d17b
Signed-off-by: Colin Cross <ccross@android.com>