CHROMIUM: ath9k: stop btcoex on device suspend
When the device is moving to suspend state, it will be powered
off and move into FULLSLEEP state. As btcoex is never be stopped,
the timer is running and try to access hw on fullsleep state.
Fix that.

Cc: stable@vger.kernel.org
Signed-off-by: Rajkumar Manoharan <rmanohar@qca.qualcomm.com>
Signed-off-by: Paul Stewart <pstew@chromium.org>

BUG=chrome-os-partner:12426
TEST=Suspend while downloading a kernel from kernel.org, resume
from suspend, confirm wifi still working, and disable and re-enable
WiFi.

Change-Id: I436127b39ac4cba9a4c056b4bb37189b8dad0c6a
Reviewed-on: https://gerrit.chromium.org/gerrit/29800
Reviewed-by: Benson Leung <bleung@chromium.org>
Commit-Ready: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>