video: s3c-fb: refactor enable/disable
Move suspend/resume work into enable/disable, so that clocks etc. are
disabled when the screen is off.  Windows are cleared on enable and
reactivated on pan.

To eliminate races with the update_regs worker, it's flushed before
disabling the display.  No further buffers are queued until the
display is enabled again.

Change-Id: I6ba4f148f6f6883f3ccf9d239aedcf9fa9423f78
Signed-off-by: Greg Hackmann <ghackmann@google.com>