spi: Fix the bug when CONFIG_DEBUG_ATOMIC_SLEEP is enabled.
When CONFIG_DEBUG_ATOMIC_SLEEP is enabled, the BUG message is shown.
It is due to the use of MUTEX in spin lock irq routine in spi.
The 'spi_pump_message function in spi.c uses spin lock irq.
In this routine, spi-s3c64xx calls dma release function which uses mutex.
This results in the BUG message like this ;
BUG: sleeping function called from invalid context at kernel/mutex.c:85
Therefore, spin lock irq routine is avoided to solve it.

Change-Id: Iaba0f8459b53a3a18c5bb2aeac987b1b67136e26
Signed-off-by: HoJune Byun <hjune.byun@samsung.com>