regulator: max77686: Add support for various operating modes
MAX77686 supports a low power mode as well as a standby mode for
its LDOs and BUCKs. Initially add LDO8 standby mode as it leaks
~160mW through a suspend/resume cycle. The idea would be to set
these modes for the rest of the LDOs and BUCKs as well.
Note: The device tree bindings have been taken from the upstream
MFD driver with the new optional bindings for operating mode and
buck ramp delay added.

BUG=chrome-os-partner:12522
TEST=Verify 1350_ldoin power during suspend along with follow-on
patches.

Change-Id: I1144254ebf6cf096bc5b11217c3045aff41d8db0
Signed-off-by: Abhilash Kesavan <a.kesavan@samsung.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/30257
Reviewed-by: Todd Broch <tbroch@chromium.org>