ARM: vexpress: Fix lock mismatch in idle_debug_set
The code path for clearing the cpumask in idle_debug_set by writing 0xff
fails to balance the cpuidle_pause_and_lock() with an unlock.

Fix this by removing the early return statement and let it fall through
to the function end by adding an else clause. At the same time, we
remove the redundant 'for' loop to keep the code tidier.

Signed-off-by: Jon Medhurst <tixy@linaro.org>