cpufreq: exynos: use range when setting cpu voltage
If the exact cpu voltage specified in the voltage table is not
supported by the regulator, regulator_set_voltage will return an
error, which is then ignored.  Request a voltage between the
required cpu voltage and the maximum cpu voltage to allow the
regulator to pick the lowest supported voltage higher than the
required voltage.  Also check the return value for errors and
abort the frequency change if necessary.

Change-Id: I8eb32ba62a95bceabf5b99d3d3b870f6e6867c73
Signed-off-by: Colin Cross <ccross@android.com>