remoteproc: add support to handle device specific resource types
The remoteproc framework handles a fixed set of resource table entries
today. To make it scalable across multiple platforms, it makes sense
for the framework to provide a way for the device specific implementation
to define and handle custom resource types. These resource types would
be very specific to a implementation instance that it does not make sense
for the framework to handle it.

For instance, a remoteproc implementation might want timers information
embedded in the resource table so that the driver could parse the binary
and enable accordingly. Another example would be hwspinlocks that it
is using, to properly share system wide resources.

This patch adds a function pointer to the list of rproc_ops for the
driver implementation to handle such custom resources.

Change-Id: I369b56da4febf1eee9ef8b8a02b6bd3e600c0324
Signed-off-by: Subramaniam Chanderashekarapuram <subramaniam.ca@ti.com>