CHROMIUM: iio: isl29018: Support suspend and resume.
The driver leaves the device in power-down state anyway,
so there is nothing to do on suspend.
On resume, we just have to make sure the range and ADC
values are updated in the device since it may have been
powered down in suspend.

BUG=chrome-os-partner:9494
TEST=Verify that lux values are generally unchanged after suspend/resume.
This is a bit difficult to reproduce on snow because we have to simulate
a FET5 power-down on suspend by manually hitting the TPS.  We also have
to modify a range or resolution value that gets pushed to the device, and
we do not normally do that.  This can reproduce the problem:
cd /sys/bus/iio/devices/*
cat in_illuminance0_input
echo 4000 > range
cat in_illuminance0_input # The value changes just a little bit.
i2cset -f -y 4 0x48 0x13 0x02 # Turn off FET5
i2cset -f -y 4 0x48 0x13 0x1f # Turn FET5 back on
powerd_suspend
cat in_illuminance0_input # The value is 4x expected without this patch.

Change-Id: I73e9f9357db3ae3c240a6e1cc5c8acbcf97b6971
Signed-off-by: Bryan Freed <bfreed@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/31662
Reviewed-by: Benson Leung <bleung@chromium.org>