CHROMIUM: Increase i2c arbitration wait time
We currently intend to wait 50ms for the EC to release the i2c bus
(although due to a bug soon to be fixed, we wait longer).

The longest transaction we expect the EC to perform is 4 bytes, which
should be about 0.5ms, but if there is a bug in the EC driver the time
could wind out significantly.

Change the wait time to 500ms.

BUG=chrome-os-partner:12420
BRANCH=snow
TEST=manual
On kernel:

while true; do i2cdump  -f -y 4 0x48; done >/dev/null

On EC:

pmu 1000

See that there are no errors.

Change-Id: I2ab5c6f0ac1182485ac1a4f60e4e76ba83fb3987
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/31923
Reviewed-by: Doug Anderson <dianders@chromium.org>