watchdog: s3c2410: Avoid excessive petting during cpufreq
On more modern Samsung SoCs (like exynos5) the watchdog is
parented on a clock that doesn't change.  Avoid all of the extra
petting, disabling, and re-enabling of the watchdog in the case
that the clock never changes.

NOTE: As pointed out in
  <https://gerrit.chromium.org/gerrit/#/c/31465/>

...the proper way would be to register a notifier for the clock
directly, but CONFIG_COMMON_CLK isn't enabled.

BUG=chrome-os-partner:9403
TEST=If I open /dev/watchdog now it eventually dies.

Change-Id: I4f982dc85caf091444633aff94b90d524bdfe829
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/32566
Reviewed-by: Olof Johansson <olofj@chromium.org>
Commit-Ready: Grant Grundler <grundler@chromium.org>
Reviewed-by: Grant Grundler <grundler@chromium.org>
Tested-by: Grant Grundler <grundler@chromium.org>