ARM: EXYNOS: Fix watchdog timer disabled after wakeup from AFTR.
The cpu use to the irom code when wakeup from AFTR and hot-plugged out.
The irom code is write the zero to the WTCON register.
Because of it, a watchdog timer disabled after cpu off and on (like to AFTR, LPA
and hot-plugged).
So, should be save and restore the WTCON register when cpu off and on.

Change-Id: I78bc07fbaf4bbc8e74dc779a1e289f6148237076
Signed-off-by: Taekki Kim <taekki.kim@samsung.com>