CHROMIUM: Refactor chromeos_ec_i2c to use a common file
Much of the code in the I2C implementation should actually be common
to all interfaces. As a first step, create a chromeos_ec.c which
implements this common functionality, and make the I2C implementation
use it.

(Note: The FDT change is in this commit also to avoid a breakage)

BUG=chrome-os-partner:13860
TEST=manual
See that snow keyboard still works correctly over I2C.

Change-Id: Ib48009d3c38ee8b80b3dd40eadbb7f56f5b5ea59
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/33571
Reviewed-by: Vincent Palatin <vpalatin@chromium.org>