KDS: make kds_resource_set_release callable from atomic context
This allows calling kds_resource_set_release() from an atomic context.
In drm/exynos, we want to release the kds resource from the page_flip
interrupt handler. Without this change, we cannot call the release
method directly and would need to instead use a work item. To avoid
this, let's use a spinlock instead of a mutex.

This change also make kds consistent with the intended semantics
of the direct parameter of kds_callback_init. The direct parameter
requires that the callback be callable from an atomic context.
Having kds_resource_set_release() NOT callable from an atomic context
defeats the purpose of direct.

BUG=chrome-os-partner:12170
TEST=VT switch. Log in/out. Suspend/resume. Web GL. Youtube.

Change-Id: I3578e2324e1f04dc6a41b60055b3ac61d3e00a52
Signed-off-by: Mandeep Singh Baines <msb@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/33403
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>